feat: add comment-incentive command
