+++ src/handlers/comment/commands.ts
@@ -7,7 +7,7 @@ export enum IssueCommentCommands {
   MULTIPLIER = "/multiplier", // set bounty multiplier (for treasury)
   QUERY = "/query",
   ASK = "/ask", // ask GPT a question
-  INCENTIVIZE = "/incentivize",
+  COMMENTINCENTIVE = "/comment-incentive",
   // Access Controls
 
   ALLOW = "/allow",
